Atualização de arquitetura
