+++ app/src/main/java/br/com/zup/projetofinalzup/data/datasource/remote/MenuItemAPI.kt
@@ -1,7 +1,7 @@
-package br.com.zup.projetofinalzup.domain.repository
+package br.com.zup.projetofinalzup.data.datasource.remote
 
-import br.com.zup.projetofinalzup.data.datasource.model.MenuItem
-import br.com.zup.projetofinalzup.domain.repository.model.MenuRequest
+import br.com.zup.projetofinalzup.data.model.MenuItem
+import br.com.zup.projetofinalzup.data.datasource.remote.model.MenuRequest
 import retrofit2.http.Body
 import retrofit2.http.Headers
 import retrofit2.http.POST
