+++ app/src/main/java/br/com/zup/projetofinalzup/data/datasource/remote/model/MenuRequest.kt
@@ -1,0 +1,5 @@
+package br.com.zup.projetofinalzup.data.datasource.remote.model
+
+class MenuRequest(
+    val document:String
+)
